Extract images from the MNIST files
The MNIST file which contains the images to be classified are obtained from this link:
https://www.kaggle.com/datasets/hojjatk/mnist-dataset?resource=download.

The extracted images are not included in the commit as they are significantly large. To
extract the images, run extract_data.py on your local repository.

A .gitignore file was also included to avoid pushing the extracted images from my local
repository to the remote repository.