Conduct Data Preprocessing
The extracted data must be preprocessed before it can be used to train the model.

The PyTorch CNN model I plan to create allows for the use of Data Loaders which are more computationally efficient.
In order to use them, we have to preprocess the data and assign them to Data Loaders.

The extracted data is split into training and test data and assigned to a training and testing Data Loader respectively.

The training-test split ensures that we have a validation set to test the performance of our data, before running it on
the test set.

A requirements.txt file has also been added to track the dependencies required for reproduction purposes.