Debug the DataLoaders
Currently, the way the data is being preprocessed results in the images and labels in the dataloaders
to be misaligned.

This results in the model learning based on incorrect data, which results in inaccurate predictions.

In the custom dataset class, the images are sorted to ensure that the images align with the labels.
This corroborates with the assumption that the data was also extracted in such an order.

Ensuring that the images in the data loader are ordered in the same manner as those in the
directories storing the images ensures that the images and labels are aligned correctly.