Train models using the data
A simple and a more complex model were trained using the the training test set, which was
further split into a training set and a validation set. The trained model was then tested on a
testing dataset to determine its performance.

Additionally, .pth files were created to store the data loaders so that preprocessing is not repeated.
Same for the optimal parameters for each trained model.