Conduct Hyperparameter Tuning
Currently, the simple CNN model is found to run in less than half the time compared to the complex CNN model.
In order to save computational time, hyperparameter tuning is done on the simple CNN model. The two hyperparameters
to be tuned are the dropout rate and learning rate.

Hyperparameter Tuning is done to find the combination of hyperparameters that result in the best performance of the model.

Code to conduct hyperparameter tuning can be found in hyperparameter_tuning.py. The best parameters are stored in
best_simple_model_weights.pth.

Hyperparameter Tuning is automated instead of being done manually to maintain efficiency.