Conduct data exploration
After extracting the data, we have to conduct some analysis on the data to identify any preprocessing steps
we have to implement.

Without data exploration, we are unable to identify and patterns or outliers in the data, which could
significantly affect the performance of our model.

The data exploration steps are stored in data-exploration.ipynb, and details steps such as identifying
missing values, outliers, and feature selection

The above-mentioned considerations are important can significantly improve the accuracy
of our model.