Conduct Data Preprocessing
Currently, the data is not ready to be used as input to the model as it has not been preprocessed.

The data must be preprocessed so that the model can train based on the correct features and target variable.

The data is split into its correlated features, and its target variable, Adjusted Closing Price. The features
are then scaled to improve convergence, and then stored.