Train models using the data
A linear regression model and a neural network model were trained using the the training test set, which was
further split into a training set and a validation set. The trained model was then tested on a
testing dataset to determine its performance.

The Linear Regression model performed better than the neural network model. Hence, no hyperparameter
tuning was done.

Files were also created to store the model parameters so they do not have to be retrained during model testing.