Conduct data exploration
The dataset used consists of Amazon product reviews. The link to the product reviews dataset can be found here: https://www.kaggle.com/datasets/gauravduttakiit/amazon-product-reviews?resource=download

The dataset must be explored to provide a better idea of the structure of the dataset, as well as how to preprocess the data appropriately.

Information about the dataset, such as the number of entries, attributes recorded, identifying variables of interest, and rating imbalances are explored in data-exploration.ipynb.

This information will help us to preprocess the data optimally to improve the performance of the model.