Conduct data preprocessing
The current dataset is not suitable for training as it has some redundant rows and is quite large, leading to computational inefficiency.

This could result in long model training durations.

The dataset was preprocessed by keeping only the relevant columns, and removing rows with missing values or 3 star ratings.

This results in a much smaller dataset which is suitable for model training.