Conduct model training
Currently, the dataset is generated via random upsampling. I thought that the model might not be able to recognise
negative examples sufficiently due to the low proportion of negative examples.

The low count of negative examples could result in the model not being able to classify unseen negative examples
accurately.

Hence, instead of only conducting random upsampling, I decided to use SMOTE analysis to upsample the dataset
as well, and see which dataset each model performs better under. As the dataset contained textual data and required
assignment to classes, I decided to train a Logistic Regression model, an RNN model, a Naive Bayes Model, and an SVM.

After comparing all models and datasets, the RNN model with randomly upsampled dataset performed the best, and will
be used for hyperparameter tuning.