Conduct hyperparameter tuning
It has been found that the RNN model produces the best results out of the four model types.
Hence, I decided to conduct hyperparameter tuning on the RNN model.

The optimal hyperparameters can be found in models.py. The accuracy and precision scores
for each class were 0.99 for all test statistics.