Implement UI
Currently, there is no user interface, which makes it inconvenient for users and
not user-friendly.

HTML files were used to allow users to host locally. The model with the best
performance obtained from hyperparameter tuning was also used for prediction.

This provides a simple interface for users to observe the effectiveness of the model.